requirements.txt
langchain
openai
gradio
faiss-cpu
sentence-transformers
PyMuPDF
tiktoken
python-dotenv
